init with login

+++ src/Hotel.java
@@ -10,7 +10,7 @@ public class Hotel {
     //방번호-방(객체)
     private Map<String, Room> roomMap;
     //보유 자산
-    private int asset;
+    private long asset;
     public Hotel() {
         roomMap = new HashMap<>();
         asset = 0;
